+++ srsenb/src/phy/txrx.cc
@@ -98,8 +98,8 @@ void txrx::run_thread()
 
   // Set Tx/Rx frequencies
   for (uint32_t cc_idx = 0; cc_idx < worker_com->get_nof_carriers(); cc_idx++) {
-    float    tx_freq_hz = worker_com->get_dl_freq_hz(cc_idx);
-    float    rx_freq_hz = worker_com->get_ul_freq_hz(cc_idx);
+    double   tx_freq_hz = worker_com->get_dl_freq_hz(cc_idx);
+    double   rx_freq_hz = worker_com->get_ul_freq_hz(cc_idx);
     uint32_t rf_port    = worker_com->get_rf_port(cc_idx);
     log_h->console(
         "Setting frequency: DL=%.1f Mhz, UL=%.1f MHz for cc_idx=%d\n", tx_freq_hz / 1e6f, rx_freq_hz / 1e6f, cc_idx);
